+++ Exodus/Changes.md
@@ -1,7 +1,9 @@
 # Utopia Exodus - List of Changes
 
 ## v4.1.7 (in development)
-Built with Scala v2.13.14
+### Other changes
+- Built with Scala v2.13.14
+- Supports Vault v1.20
 
 ## v4.1.6 - 28.07.2024
 Conforms to changes introduced in **Flow v2.4** and **Vault v1.19**