+++ Logos/Changes.md
@@ -1,7 +1,9 @@
 # Utopia Logos - List of Changes
 
 ## v0.2.1 (in development)
-Built with Scala v2.13.14
+### Other changes
+- Built with Scala v2.13.14
+- Supports Vault v1.20
 
 ## v0.2 - 28.07.2024
 The first beta release. Expect breaking changes in future updates.